+++ README.md
@@ -25,7 +25,7 @@ Read the documentation and sources of each module how to use it in the project.
 - [http](./src/http/README.md)
 - [logger](./src/logger/README.md)
 - [processing](./src/processing/README.md)
-- [promises](./src/promises/README.md)
+- [promise-utils](./src/promise-utils/README.md)
 - [release-scripts](./src/release-scripts/README.md)
 - [run-in-loop](./src/run-in-loop/README.md)
 - [utils](./src/utils/README.md)
