+++ src/release-scripts/tag-and-release.ts
@@ -19,7 +19,7 @@ import { readFileSync } from 'node:fs';
 
 import { Octokit } from '@octokit/rest';
 
-import { go } from '../promises';
+import { go } from '../promise-utils';
 
 const execSyncWithErrorHandling = (command: string) => {
   // eslint-disable-next-line functional/no-try-statements
