+++ src/universal-index.ts
@@ -1,6 +1,6 @@
 // NOTE: Only export modules which work in both Node.js and browser environments.
 export * from './blockchain-utilities';
 export * from './http';
-export * from './promises';
+export * from './promise-utils';
 export * from './run-in-loop';
 export * from './utils';
